telegram-bot/requirements.txt
# dLNk Telegram Bot Dependencies
# Python 3.9+ required

# Telegram Bot Framework (aiogram 3.x)
aiogram>=3.3.0

# HTTP Client
aiohttp>=3.9.0
httpx>=0.26.0

# Environment Variables
python-dotenv>=1.0.0

# Async Scheduler
apscheduler>=3.10.0

# Database (optional, for local caching)
aiosqlite>=0.19.0

# Validation
pydantic>=2.5.0

# Logging
structlog>=24.1.0

# Rate Limiting
aiolimiter>=1.1.0

# Cryptography (for license verification)
cryptography>=41.0.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.23.0
